requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.5
django-allauth==0.42.0
django-cors-headers==3.5.0
django-extensions==3.0.9
django-rest-auth==0.9.5
djangorestframework==3.11.1
djongo==1.3.3
dnspython==2.0.0
gunicorn==20.0.4
idna==2.10
oauthlib==3.1.0
pymongo==3.11.0
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.2.4
urllib3==1.25.10
whitenoise==5.2.0
